# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "在家办公小程序接口文档",
        "version": "1.0"
    },
    "paths": {
...
        "/api/zaochenjiayourecord/save": {
            "post": {
                "tags": [
                    "BetterMe"
                ],
                "summary": "早晨加油",
                "operationId": "App\\Http\\Controllers\\Api\\ZaoChenJiaYouRecordController::save",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "contents": {
                                        "description": "内容",
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "description": "用户id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}